restructuring code tree following Olivia's suggestions
