+++ src/rail/examples_data/creation_data/data/fsps_default_data/test_fsps_sed.fits
